mudanças nos controllers

+++ Entrega-Atividade-Full-Stack-main/controllers/task_controller.py
@@ -1,4 +1,4 @@
-from flask import render_template,redirect,url_for,jsonify,request
+from flask import render_template,redirect,url_for,request
 from models.task import Task
 from models.user import User
 from config import db
@@ -13,27 +13,32 @@ class TaskController:
     def create_task():
            if request.method == 'POST':
               data = request.form
-              task = Task(title=data['title'],description=data['description'],status=data['status'],user_id=data['user_id'])
+              task = Task(id=data['id'],title=data['title'],description=data['description'],status=data['status'],user_id=data['user_id'])
               db.session.add(task)
               db.session.commit()          
               return redirect(url_for('list_tasks'))
            
            users = User.query.all()
-           return render_template('new_task.html',users=users)
+           return render_template('create_task.html',users=users)
+    
     @staticmethod
     def update_task_status(id):
             task = Task.query.get(id)
             if task:
-                  task.status = 'concluído'                  
+                  task.status = 'Concluído'                 
                   db.session.commit()
                   return redirect(url_for('list_tasks'))
-            return render_template('tasks.html')
+            
+            tasks = Task.query.all()
+            return render_template('tasks.html',tasks=tasks)
            
     @staticmethod
     def delete_task(id):
             task = Task.query.get(id)
-            db.session.delete(task)
-            db.session.commit()
-            return redirect(url_for('list_tasks') )
+            if task:
+                db.session.delete(task)
+                db.session.commit()
+                return redirect(url_for('list_tasks') )
+            return render_template('tasks.html')
